+++ src/contexts/ComponentAdapter/componentAdapterTypes.ts
@@ -2,6 +2,7 @@ export type { ProgressBarProps } from '@/components/Common/UI/ProgressBar/Progre
 export type { BreadcrumbsProps } from '@/components/Common/UI/Breadcrumbs/BreadcrumbsTypes'
 export type { ButtonProps, ButtonIconProps } from '@/components/Common/UI/Button/ButtonTypes'
 export type { CardProps } from '@/components/Common/UI/Card/CardTypes'
+export type { BoxProps } from '@/components/Common/UI/Box/BoxTypes'
 export type { CheckboxProps } from '@/components/Common/UI/Checkbox/CheckboxTypes'
 export type {
   CheckboxGroupProps,